// src/app/models/staff/staff-attendance-add-result-view.ts
import { StaffAttendanceAddView } from './staff-attendance-add-view';

/**
 * Teacher API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



/**
 * StaffAttendanceAddResultView
 */
export interface StaffAttendanceAddResultView {
    /**
     * Gets or Sets StaffAttendancAddViews
     */
    staffAttendancAddViews?: Array<StaffAttendanceAddView>;
    /**
  * Gets or Sets IsAbsent
  */
    isAbscent?: string;
    /**
     * Gets or Sets InTime
     */
    inTime?: string;
    /**
     * Gets or Sets IsAbsent
     */
    outTime?: string;
    /**
    * Gets or Sets Calendar_Date
    */
    calendarDate?: string;
    /**
     * 
     */
    selectedOperation?: any;
}


    